+++ tests/testthat/test-provider-databricks.R
@@ -124,3 +124,15 @@ test_that("the user agent respects SPARK_CONNECT_USER_AGENT when set", {
     expect_match(databricks_user_agent(), "^testing r-ellmer")
   )
 })
+
+test_that("tokens can be requested from a Connect server", {
+  skip_if_not_installed("connectcreds")
+
+  withr::local_envvar(
+    DATABRICKS_HOST = "https://example.cloud.databricks.com",
+    DATABRICKS_TOKEN = "token"
+  )
+  connectcreds::local_mocked_connect_responses(token = "token")
+  credentials <- default_databricks_credentials()
+  expect_equal(credentials(), list(Authorization = "Bearer token"))
+})